refactor(ai-agent,conversation): move session-scoped ACP ops to conversation domain
Session-scoped endpoints under /api/conversations/{id}/* (mode, model,
usage, slash-commands, side-question, workspace, openclaw/runtime) now
live in aionui-conversation instead of aionui-ai-agent. Motivation: the
URL namespace belongs to the conversation domain, and keeping these
handlers in ai-agent made the two crates' responsibilities fuzzy.

Design: ai-agent exposes dispatch methods on AgentInstance only;
ConversationService forwards through task(id) and never touches the
concrete agent variants. aionui-conversation no longer depends on
agent-client-protocol — SDK types reach it via AgentInstance returns.

Dropped endpoints (confirmed unused by the frontend):
- /api/acp/{env,detect-cli,health-check,probe-model}
- /api/conversations/{id}/config[/{configId}]
- /api/conversations/{id}/agent-capabilities
- /api/conversations/{id}/reload-context

Added /api/agents/health-check to replace /api/acp/health-check.
Renamed /api/conversations/{id}/stop to /cancel.

get_usage now returns serde_json::Value with snake_case key
normalization (via aionui_common::normalize_keys_to_snake_case),
instead of the typed SDK UsageUpdate — matches the AionUI wire
convention and decouples the frontend from the SDK schema.

AppConfig.data_dir / AppServices.data_dir migrated from String to
PathBuf; call sites drop the PathBuf::from(&services.data_dir)
wrapper.
