+++ crates/aionui-ai-agent/src/services/mod.rs
@@ -1,6 +1,6 @@
+pub mod agent;
 pub mod custom;
-pub mod openclaw;
 pub mod remote;
-pub mod service;
+
+pub use agent::AgentService;
 pub use remote::RemoteAgentService;
-pub use service::AgentService;
