+++ crates/aionui-conversation/src/lib.rs
@@ -2,7 +2,9 @@
 mod convert;
 pub mod response_middleware;
 pub mod routes;
+pub mod routes_aux;
 pub mod service;
+mod service_ops;
 pub mod skill_resolver;
 pub mod skill_snapshot;
 pub mod state;
@@ -13,6 +15,7 @@ pub use response_middleware::{
     MiddlewareResult, detect_cron_commands, has_cron_commands, strip_cron_commands, strip_think_tags,
 };
 pub use routes::conversation_routes;
+pub use routes_aux::conversation_ops_routes;
 pub use service::{ConversationService, OnConversationDelete};
 pub use state::ConversationRouterState;
