+++ crates/aionui-conversation/src/state.rs
@@ -6,6 +6,6 @@ use aionui_ai_agent::IWorkerTaskManager;
 /// Shared state for conversation route handlers.
 #[derive(Clone)]
 pub struct ConversationRouterState {
-    pub conversation_service: ConversationService,
-    pub worker_task_manager: Arc<dyn IWorkerTaskManager>,
+    pub service: ConversationService,
+    pub task_manager: Arc<dyn IWorkerTaskManager>,
 }
