+++ crates/aionui-cron/src/executor.rs
@@ -2106,6 +2106,7 @@ mod tests {
             std::env::temp_dir(),
             stub_broadcaster,
             Arc::new(StubSkillResolver),
+            Arc::new(StubTaskManager),
             Arc::clone(&stub_repo),
             Arc::clone(&agent_metadata_repo),
             acp_session_repo,
@@ -2709,6 +2710,7 @@ mod tests {
             std::env::temp_dir(),
             Arc::clone(&broadcaster),
             Arc::new(StubSkillResolver),
+            Arc::clone(&task_manager),
             Arc::clone(&repo),
             Arc::clone(&agent_metadata_repo),
             acp_session_repo,
